Create superclass web_handler.dart
